examples/snpe/inception_snpe_qualcomm_npu/requirements.txt
pillow
tabulate
torchvision
